Created new key/cert that doesn't contain port in CN
This allows the same cert to be used for both cas and app server
in the tests. But, it causes problems when the port is included,
but doesn't match what the server is using.

Create rootCA that can sign the request, allowing future work
to specify the CA to trust when communicating with the CAS server
for validations